Change runner to extend all timeouts to 90s
This change also causes ginkgo to display when commands finish along with a report of how long they took. We did this to aid analysis of tests that were timing out and allow cats to pass as a purely functional test (with no performance testing aspect) on slower CF instances.

Signed-off-by: Jesse Alford <jalford@pivotallabs.com>